mnn_llm_benchmark/pyproject.toml
[project]
name = "mnn-llm-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pathlib2>=2.3.7.post1",
    "plotly>=6.4.0",
    "psutil>=7.1.3",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "toml>=0.10.2",
]
